\name{C19dNUTS-package}
\alias{C19dNUTS-package}
\alias{C19dNUTS}
\docType{package}
\title{
Dataset of Regional COVID-19 Deaths per 100,000 Pop (NUTS)
}
\description{
The package contains a dataset with cumulative COVID-19 deaths (absolute and per 100,000 pop) at the regional level (mostly NUTS 3) for 31 EU/EFTA countries (AT, BE, BG, CH, CY, CZ, DE, DK, EE, EL, ES, FI, FR, HR, HU, IE, IS, IT, LT, LU, LV, MT, NL, NO, PL, PT, RO, SE, SI, SK, UK). The data enable a regional or spatial analysis of COVID-19 deaths in Europe at a small scale level. The methodological basis for such statistical analyzes can be found in geographical/epidemiological works on COVID-19 in individual countries (e.g., Bourdin et al. 2021, Mitze & Kosfeld 2021, Saffary et al. 2020, Wieland 2020, 2022).

Please note that this dataset, unlike the \emph{COVID-19 European regional tracker} (Naqvi 2021), contains only cumulative numbers and no time series. This is because many countries only publish cumulative data on COVID-19 deaths, while reported COVID-19 cases are often published in the form of daily data. The data is as of June 2022. Please look at the description of the dataset set if you want to get more detailed information about technical details or limitations of the data. 
}

\author{
Thomas Wieland

Maintainer: Thomas Wieland \email{thomas.wieland.kit@googlemail.com}
}

\references{
Bourdin, S., Jeanne, L., Nadou, F. & Noiret, G. (2021): \dQuote{Does lockdown work? A spatial analysis of the spread and concentration of Covid-19 in Italy}. In: \emph{Regional Studies}, \bold{55}, p. 1182-1193. \doi{10.1080/00343404.2021.1887471}.

Mitze, T. & Kosfeld, R. (2021): \dQuote{The propagation effect of commuting to work in the spatial transmission of COVID-19}. In: \emph{Journal of Geographical Systems}, \bold{24}, p. 5-31. \doi{10.1007/s10109-021-00349-3}.

Naqvi, A. (2021): \dQuote{COVID-19 European regional tracker}. In: \emph{Scientific Data}, \bold{8}, 181 . \doi{10.1038/s41597-021-00950-7}.

Saffary, T., Adegboye, O.A., Gayawan, E., Elfaki, F., Kuddus, M.A. & Saffary, R. (2020): \dQuote{Analysis of COVID-19 Cases' Spatial Dependence in US Counties Reveals Health Inequalities}. In: \emph{Frontiers in Public Health}, \bold{8}. \doi{10.3389/fpubh.2020.579190}.

Wang, Q., Dong, W., Yang, K., Ren, Z., Huang, D., Zhang, P. & Wang, J. (2021): \dQuote{Temporal and spatial analysis of COVID-19 transmission in China and its influencing factors}. In: \emph{International Journal of Infectious Diseases}, \bold{105}, p. 675-685. \doi{10.1016/j.ijid.2021.03.014}.

Wieland, T. (2020): \dQuote{Flatten the Curve! Modeling SARS-CoV-2/COVID-19 Growth in Germany at the County Level}. In: \emph{REGION}, \bold{7}, 2, p. 43-83. \doi{10.18335/region.v7i2.324}.

Wieland, T. (2022): \dQuote{Spatial patterns of excess mortality in the first year of the COVID19 pandemic in Germany}. In: \emph{European Journal of Geography}, \bold{13}, 4, p. 18-33. \doi{10.48088/ejg.t.wie.13.4.018.033}.
}